// app/shared/dashboard/widget/constants/chart-type.enum.ts
export enum ChartType {
    LINE_CHART = 0,

    PIE_CHART_AVG_TOP5 = 14,
    PIE_CHART_LAST_TOP5 = 13,
    PIE_CHART_LAST_TOP10 = 22,
    PIE_CHART_AVG_TOP10 = 23,
    PIE_CHART_LAST_ALL = 24,
    PIE_CHART_AVG_ALL = 25,
    PIE_CHART_LAST_BOTTOM10 = 26,
    PIE_CHART_AVG_BOTTOM10 = 27,

    BAR_CHART_AVG_ALL = 28,
    BAR_CHART_AVG_TOP5 = 29,
    BAR_CHART_AVG_TOP10 = 30,
    BAR_CHART_AVG_BOTTOM10 = 31,

    AREA_CHART = 6,
    DUAL_LINE_BAR = 4,

    STACKED_AREA_CHART = 12,
    STACKED_BAR_CHART = 11,

    PERCENTILE_GRAPH = 7,
    SLAB_COUNT_GRAPH = 8,

    DIAL_CHART_LAST = 18,
    DIAL_CHART_AVG = 17,

    METER_CHART_LAST = 20,
    METER_CHART_AVG = 19,

    DUAL_AXIS_LINE = 3,

    DONUT_CHART_LAST = 32,
    DONUT_CHART_AVG = 33,

     GEO_MAP_AVG = 16,
    GEO_MAP_LAST = 15,
    GEO_MAP_MAX = 67,
    GEO_MAP_MAX_OF_AVG = 84,

    GEO_MAP_EXTENDED_AVG = 65,
    GEO_MAP_EXTENDED_LAST = 64,
    GEO_MAP_EXTENDED_MAX = 66,

    LINE_STACKED = 5,
    MIN_MAX_GRAPH = 21,

    PERCENTILE_BAR_CHART = 59,
    PERCENTILE_PIE_CHART = 60,
    PERCENTILE_DUAL_LINE_BAR = 36,
    PERCENTILE_STACKED_AREA_CHART = 37,
    PERCENTILE_STACKED_BAR_CHART = 38,
    PERCENTILE_DIAL_CHART = 39,
    PERCENTILE_METER_CHART = 40,
    PERCENTILE_DUAL_AXIS_LINE = 41,
    PERCENTILE_DONUT_CHART = 42,
    PERCENTILE_LINE_STACKED_BAR = 43,
    PERCENTILE_AREA_CHART = 44,

    SLAB_COUNT_LINE_CHART = 45,
    SLAB_COUNT_PIE_CHART = 46,
    SLAB_COUNT_DUAL_LINE_BAR = 47,
    SLAB_COUNT_STACKED_AREA_CHART = 48,
    SLAB_COUNT_STACKED_BAR_CHART = 49,
    SLAB_COUNT_DIAL_CHART = 50,
    SLAB_COUNT_METER_CHART = 51,
    SLAB_COUNT_DUAL_AXIS_LINE = 52,
    SLAB_COUNT_DONUT_CHART = 53,
    SLAB_COUNT_LINE_STACKED_BAR = 54,
    SLAB_COUNT_AREA_CHART = 55,

    GRAPH_STATS_AVG = 0,
    GRAPH_STATS_MIN = 1,
    GRAPH_STATS_MAX = 2,
    GRAPH_STATS_COUNT = 3,
    GRAPH_STATS_SUMCOUNT = 4,
    GRAPH_STATS_LAST = 5,
    GRAPH_STATS_SUM = 6,

    NORMAL_GRAPH_TYPE = 56,
    PERCENTILE_GRAPH_TYPE = 57,
    SLAB_COUNT_GRAPH_TYPE = 58,

    DUAL_AXIS_AREA_LINE = 80,

    DUAL_AXIS_STACKED_BAR = 83,
    PERCENTILE_DUAL_AXIS_STACKED_BAR = 89,
    SLAB_COUNT_DUAL_AXIS_STACKED_BAR = 90,

    DUAL_AXIS_STACKED_AREA = 88,
    PERCENTILE_DUAL_AXIS_STACKED_AREA = 91,
    SLAB_COUNT_DUAL_AXIS_STACKED_AREA = 92,

    PERCENTILE_DUAL_AXIS_AREA_LINE = 81,
    SLAB_COUNT_DUAL_AXIS_AREA_LINE = 82,

    TABULAR_GRAPH_STATUS_BASED = 0,
    TABULAR_VECTOR_BASED = 1,

    CATEGORY_STACKED_BAR_CHART = 61,
    CATEGORY_STACKED_AREA_CHART = 62,

    /** Correlated MultiAxis graph type Constant*/
    CORRELATED_MULTI_AXES_CHART = 84,
    CORRELATED_MULTI_AXES_CHART_LINE = 85,
    CORRELATED_MULTI_AXES_CHART_AREA = 86,
    CORRELATED_MULTI_AXES_CHART_BAR = 87,

    /*Charts with External Data. */
    SCATTER_CHART = 121,
    REVENUE_PREDICTION_CHART = 122
}